minor issue!

+++ w_grid.cpp
@@ -6,6 +6,6 @@ bool is_0bstacle(std::pair<short int, short int> coordinate) {
     return worker_pos[coordinate.first][coordinate.second] == 1;
 }
 
-void init_worker_grid(size_t width=65, size_t height=65) {
+void init_worker_grid(short width=65, short height=65) {
     worker_pos = vector<vector<short>>(width, vector<short>(height, 0));
 }
